chore: move mapper to separate file and create only one mapper in the graph server
