+++ router/core/graph_server.go
@@ -629,8 +629,7 @@ func (s *graphServer) buildGraphMux(ctx context.Context,
 				metricSetAttributes: b,
 				metricsEnabled:      metricsEnabled,
 				traceEnabled:        s.traceConfig.Enabled,
-				filterEnabled:       !s.metricConfig.UseCloudExporter,
-				includedAttributes:  s.metricConfig.Attributes,
+				mapper:              mapper,
 				w:                   w,
 				r:                   r,
 			})
